+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- [fj-doc-playground-quarkus] doc project init, default venus version 8.8.5
 - [fj-doc-playground-quarkus] quarkus version set to 3.14.3
 
 ## [8.8.5] - 2024-09-12